feat: redesign frontend, fix memory leak and clean backend methods
- Backend: Removed legacy 'geodesic' method; promoted PySAL as default; removed generic Exception handlers.
- Frontend: Fixed memory leak in main.py (static vs shared loading); Applied new UI/UX with Inter font and Dark Mode overrides.
- UI: Added 'Info' landing page, refined Dashboard layout, added 'Back to Top' button, and semantic insight cards.
- Docker: Optimized container build context.
