+++ app/analysis/__init__.py
@@ -1,0 +1,1 @@
+from .reporting import analyze_allocation, generate_allocation_pdf
